+++ ownCloud/Settings/SettingsViewController.swift
@@ -41,6 +41,7 @@ class SettingsViewController: StaticTableViewController {
 
 			if #available(iOS 13, *), // Require iOS 13
 			   !OCLicenseEMMProvider.isEMMVersion, // Do not show purchases in the EMM version
+			   !VendorServices.shared.isBranded,
 			   OCLicenseEnterpriseProvider.numberOfEnterpriseAccounts < OCBookmarkManager.shared.bookmarks.count { // Do only show purchases section if there's at least one non-Enterprise account
 				self.addSection(PurchasesSettingsSection(userDefaults: userDefaults))
 			}