+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # ChangeLog
 
+## Release version 11.4.2 (November 2020)
+
+- Support for new Display Sizes
+- Favorites in Directory Picker (#814)
+- Rename Filename in Scan View (#822)
+- Fix: Save Attachments from Mail.app (#816)
+- Fix: Authentication Error in FileProvider (#808)
+- Fix: New Folder in FileProvider (#827)
+
 ## Release version 11.4.1 (September 2020)
 
 - Image Metadata (#679)